PoC: Allow auto-draft in API orders
For some operations in the API, like calculating taxes or shipping, it's
helpful to be able to have an order saved in the database. Having an
order object with an ID lets us perform other API actions on that order.
However, we also don't necessarily want these to show up on certain
screens.

Here we adopt the auto-draft status from core WordPress as a way to save
an order as a discrete object while signalling that it's not ready to
action yet.

Suggestions:

* auto-draft might not make sense given that it's used in core. We might
  want to use something like `draft` instead.
* We may want to track this separate set of statuses (trash, auto-draft)
  in a more official way, so we don't have to copy/paste around the
  codebase.